Extract solana-keypair, move other things around
