+++ sdk/Cargo.toml
@@ -28,17 +28,18 @@ full = [
     "rand0-7",
     "serde_json",
     "solana-signature",
-    "solana-signer",
     "ed25519-dalek",
     "libsecp256k1",
     "sha3",
     "solana-commitment-config",
     "digest",
     "solana-pubkey/rand",
+    "dep:solana-keypair",
     "dep:solana-precompile-error",
     "dep:solana-presigner",
     "dep:solana-seed-derivable",
     "dep:solana-seed-phrase",
+    "dep:solana-signer",
     "dep:solana-transaction-error"
 ]
 borsh = ["dep:borsh", "solana-program/borsh", "solana-secp256k1-recover/borsh"]
@@ -106,6 +107,7 @@ solana-frozen-abi-macro = { workspace = true, optional = true, features = [
 ] }
 solana-inflation = { workspace = true, features = ["serde"] }
 solana-instruction = { workspace = true }
+solana-keypair = { workspace = true, optional = true }
 solana-native-token = { workspace = true }
 solana-packet = { workspace = true, features = ["bincode", "serde"] }
 solana-precompile-error = { workspace = true, optional = true }
@@ -126,7 +128,7 @@ solana-signature = { workspace = true, features = [
     "std",
     "verify",
 ], optional = true }
-solana-signer = { workspace = true, optional = true, features = ["keypair"] }
+solana-signer = { workspace = true, optional = true }
 solana-time-utils = { workspace = true }
 solana-transaction-error = { workspace = true, features = ["serde"], optional = true }
 thiserror = { workspace = true }
