+++ sdk/keypair/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "solana-signer"
-description = "Abstractions for Solana transaction signers. See `solana-keypair` for a concrete implementation."
-documentation = "https://docs.rs/solana-signer"
+name = "solana-keypair"
+description = "Concrete implementation of a Solana `Signer`."
+documentation = "https://docs.rs/solana-keypair"
 version = { workspace = true }
 authors = { workspace = true }
 repository = { workspace = true }
@@ -10,13 +10,22 @@ license = { workspace = true }
 edition = { workspace = true }
 
 [dependencies]
+bs58 = { workspace = true, features = ["std"] }
+ed25519-dalek = { workspace = true }
+rand0-7 = { workspace = true }
 solana-pubkey = { workspace = true }
+solana-seed-phrase = { workspace = true }
 solana-signature = { workspace = true }
+solana-signer = { workspace = true }
 solana-transaction-error = { workspace = true }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+wasm-bindgen = { workspace = true }
 
 [dev-dependencies]
 serde_json = { workspace = true }
 static_assertions = { workspace = true }
+tiny-bip39 = { workspace = true }
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
