+++ sdk/seed-derivable/Cargo.toml
@@ -13,8 +13,7 @@ edition = { workspace = true }
 ed25519-dalek = { workspace = true }
 ed25519-dalek-bip32 = { workspace = true }
 solana-derivation-path = { workspace = true }
-solana-seed-phrase = { workspace = true, features = ["keypair"] }
-solana-signer = { workspace = true }
+solana-keypair = { workspace = true }
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
