+++ sdk/signer/src/lib.rs
@@ -14,8 +14,6 @@ use {
     },
 };
 
-#[cfg(feature = "keypair")]
-pub mod keypair;
 pub mod null_signer;
 pub mod signers;
 
@@ -220,60 +218,4 @@ pub trait EncodableKeypair: EncodableKey {
 
     /// Returns an encodable representation of the associated public key.
     fn encodable_pubkey(&self) -> Self::Pubkey;
-}
-
-#[cfg(test)]
-mod tests {
-    use {super::*, crate::keypair::Keypair};
-
-    fn pubkeys(signers: &[&dyn Signer]) -> Vec<Pubkey> {
-        signers.iter().map(|x| x.pubkey()).collect()
-    }
-
-    #[test]
-    fn test_unique_signers() {
-        let alice = Keypair::new();
-        let bob = Keypair::new();
-        assert_eq!(
-            pubkeys(&unique_signers(vec![&alice, &bob, &alice])),
-            pubkeys(&[&alice, &bob])
-        );
-    }
-
-    #[test]
-    fn test_containers() {
-        use std::{rc::Rc, sync::Arc};
-
-        struct Foo<S: Signer> {
-            #[allow(unused)]
-            signer: S,
-        }
-
-        fn foo(_s: impl Signer) {}
-
-        let _arc_signer = Foo {
-            signer: Arc::new(Keypair::new()),
-        };
-        foo(Arc::new(Keypair::new()));
-
-        let _rc_signer = Foo {
-            signer: Rc::new(Keypair::new()),
-        };
-        foo(Rc::new(Keypair::new()));
-
-        let _ref_signer = Foo {
-            signer: &Keypair::new(),
-        };
-        foo(Keypair::new());
-
-        let _box_signer = Foo {
-            signer: Box::new(Keypair::new()),
-        };
-        foo(Box::new(Keypair::new()));
-
-        let _signer = Foo {
-            signer: Keypair::new(),
-        };
-        foo(Keypair::new());
-    }
 }
